+++ src/wear.rs
@@ -1,0 +1,18 @@
+use strum_macros::{Display, EnumString};
+use num_enum::{TryFromPrimitive, IntoPrimitive};
+use serde_repr::{Serialize_repr, Deserialize_repr};
+
+#[derive(Serialize_repr, Deserialize_repr, Debug, Hash, Eq, PartialEq, Display, EnumString, TryFromPrimitive, IntoPrimitive, Clone)]
+#[repr(u8)]
+pub enum Wear {
+    #[strum(serialize = "Factory New")]
+    FactoryNew = 1,
+    #[strum(serialize = "Minimal Wear")]
+    MinimalWear = 2,
+    #[strum(serialize = "Field-Tested")]
+    FieldTested = 3,
+    #[strum(serialize = "Well-Worn")]
+    WellWorn = 4,
+    #[strum(serialize = "Battle Scarred")]
+    BattleScarred = 5
+}